feat: rescope the form so it can consume the extension context
