+++ src/components/frame/Extensions/pages/ExtensionPage/partials/ExtensionPageForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ExtensionPageForm.tsx';
