+++ src/components/frame/Extensions/pages/ExtensionPage/partials/ExtensionPageForm/types.ts
@@ -1,0 +1,7 @@
+import { type RefetchExtensionDataFn } from '~hooks/useExtensionData.ts';
+import { type AnyExtensionData } from '~types/extensions.ts';
+
+export interface ExtensionPageFormProps {
+  extensionData: AnyExtensionData;
+  refetchExtensionData: RefetchExtensionDataFn;
+}
